% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_abstract.R
\name{au_abstract}
\alias{au_abstract}
\title{Open file to change the abstract}
\usage{
au_abstract()
}
\description{
The file abstract.html in your working AUTHOREA directory contains the text
of the abstract. The function simply opens the file and the content can be
changed. The text should be within <div> tags. Do not forget to save the file
after having adapted the abstract.
}
