% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_add_textblock.R
\name{au_add_textblock}
\alias{au_add_textblock}
\title{Add textblock to the manuscript}
\usage{
au_add_textblock(name)
}
\arguments{
\item{name}{Name of textblock to be added.}
}
\description{
With this function a new textblock is added at the end of the manuscript. If
you want to change the order use au_layout()
}
