% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{au_directory}
\alias{au_directory}
\title{Set custom name of AUTHOREA directory}
\usage{
au_directory(name)
}
\arguments{
\item{name}{Name of directory that contains the AUTHOREA blocks.}
}
\description{
Set the directory name that contains the content of blocks of the manuscript.
Defeault setting is "AUTHOREA". Best leave it like that.
}
