% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{au_image_directory}
\alias{au_image_directory}
\title{Set custom name of directory for figures}
\usage{
au_image_directory(name)
}
\arguments{
\item{name}{Name of directory that contains the AUTHOREA blocks.}
}
\description{
Set the directory name that contains the figures for the manuscript.
Defeault setting is "figures". Best leave it like that.
}
