% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_layout.R
\name{au_layout}
\alias{au_layout}
\title{Open Layout-file to change order of content}
\usage{
au_layout()
}
\description{
The layout.md is a file that is needed by AUTHOREA to know the order of the
files (called 'blocks' in AUTHOREA) that should be included into the
manuscript. The function simply opens the layout file and the order of the
blocks can then be addapted.
}
