% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{au_set_IDs}
\alias{au_set_IDs}
\title{Set user-ID and manuscript ID}
\usage{
au_set_IDs(userID, msID)
}
\arguments{
\item{userID}{ID number of of the owner of the manuscript in AUTHOREA.}

\item{msID}{ID number of of the manuscript.}
}
\description{
Set the user ID and the manuscript ID. Both number you can obtain from the
url to your manuscript in AUTHOREA.
}
