% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_title.R
\name{au_title}
\alias{au_title}
\title{Open file to change the titel}
\usage{
au_title()
}
\description{
The title.html was produced by AUTHOREA that contains the title of the
manuscript. The function simply opens the title file and the content can be
changed. The text of the titel should be within <div> tags. Do not forget to
save the file after having adapted the title.
}
